requirements.txt
aiogram==3.15.0
aiohttp==3.11.10
asyncpg==0.30.0
openai==1.59.5
python-dotenv==1.0.1
requests==2.32.3
